Add casts when storing values in a struct timespec.
Fixes -Wconversion warnings on some 32-bit systems where time_t is
still 32-bit.

+++ plugins/sudoers/defaults.c
@@ -999,7 +999,7 @@ store_timespec(const char *str, struct sudo_defs_types *def)
 		ts.tv_sec++;
 		nsec -= 1000000000;
 	    }
-	    ts.tv_nsec = nsec;
+	    ts.tv_nsec = (long)nsec;
 	}
     }
     if (sign == '-') {
